catsdogs/keras_catsdogs_deploy/requirements.txt
tensorflow-gpu==1.15
pillow
scipy
keras
wget
click
numpy
